web/off_dah_rails_m8/solve/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
mysql-connector-python==9.3.0
packaging==25.0
requests==2.32.4
urllib3==2.5.0
